Allow sharing modules dir between many bots, compile cleanly with -Wextra
The module data directory is now /data instead of /modules/data. This
allows /modules to be a symlink to a shared location while each bot can
have its own set of /data files.
This directory will be auto migrated to the new location if possible.

Added modules.include and modules.exclude files. These allow controling
which modules in the /modules dir will be loaded, instead of the
previous behaviour of always loading them all. Again, this is useful if
/modules is being shared.

Compile with -Wextra, ignoring a few things. The ignores made CFLAGS very
long, so it has been moved to the src/CFLAGS file, and included using
gcc/clang's @ option.

Also tweak a few things that were brought up by clang --analyze.

+++ src/mod_help.c
@@ -104,8 +104,8 @@ static void help_cmd(const char* chan, const char* name, const char* arg, int cm
 		// TODO: skip control char of cmds?
 
 		const char* found_cmd = NULL;
-		int found_cmd_idx = 0;
-		int found_len = 0;
+		size_t found_cmd_idx = 0;
+		size_t found_len = 0;
 
 		for(IRCModuleCtx** m = mods; *m; ++m){
 			if(!(*m)->commands) continue;
@@ -139,7 +139,7 @@ static void help_cmd(const char* chan, const char* name, const char* arg, int cm
 
 		if(found_cmd){
 			const char* help_text = found_mod->cmd_help ? found_mod->cmd_help[found_cmd_idx] : "No help available :(";
-			ctx->send_msg(chan, "[mod_%s cmd]: %.*s %s", found_mod->name, found_len, found_cmd, help_text);
+			ctx->send_msg(chan, "[mod_%s cmd]: %.*s %s", found_mod->name, (int)found_len, found_cmd, help_text);
 			return;
 		}
